% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{Documentation for FizzBuzz}
\usage{
FizzBuzz(n)
}
\arguments{
\item{a}{A real number}
}
\value{
Fizz if \code{a} is divisible by 3. Buzz if \code{a} is divisble by
5. FizzBuzz if \code{a} is divisible by 3 and 5.
}
\description{
Fizzbuzz recreates the game of it's same game in which people take turns
counting up from one, except every number that's divisible by 3 needs to be
said as Fizz. Every number that's divisible by 5 is replaced with Buzz. And
every number divisible by 3 and 5 is replaced with FizzBuzz
}
\details{
This is still in the description part of the documentation and and it
will be until we see something that indicates a new section.
}
\examples{
FizzBuzz(12)
FizzBuzz(5)
FizzBuzz(6)
FizzBuzz(30)
}
