% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{Documentation for Flagstaff Weather Package}
\format{
A data frame with 25449 observations with 5 columns.
\describe{
   \item{Date}{The date that the data was observed on from each of the
               subsequent columns}
   \item{PRCP}{Amount of precipitation in inches observed on that date}
   \item{Snowfall}{Amount of snowfall in inches observed on that date}
   \item{TMAX}{The maximum temperature observed on that date in fahrenheit}
   \item{TMIN}{The minimum temperature observed on that date in fahrenheit}
}
}
\usage{
Flagstaff_Weather
}
\description{
A this package contains a data set of weather information from Flagstaff's
Pulliam Airport from 1950 to 2019. Precipitation, snowfall, minimum
temperature and maximum temperature will be included for each date.
}
\keyword{datasets}
